Re #26997 Convert NormToSANSMonitor from class to func
Converts NormalizeToSANSMonitor from a class to a function as we were
not storing any persistant state within it
